Display quote/message in colour, quote author in black
In addition to the quote retrieved from the They Said So REST API,
the author is now retrieved as well. In 'quote' mode, the quote
(or override message if INKY_MESSAGE is set) is printed in "colour",
which is red, yellow or grey, depending on the Inkyphat used. A black
border is printed around the message, with the quote author (if any)
printed in black in the footer. The text colour can be reverted to
black-only by setting the BLACK_ONLY environment variable.

Change-type: minor
Signed-off-by: AlidaOdendaal <alida@balena.io>